Separating the Answer logic
